refactor: move Ref and FurtherReading components out of academy scope
